Move IPAddressSedes to common location
